txt file datasets

+++ CustomData/customdata_ori.py
@@ -60,27 +60,6 @@ class RoiDataset(Dataset):
         with open(image_set_file) as f:
             image_index = [x.strip() for x in f.readlines()]
         self.image_index = image_index
-        # ========================================================================
-        # with open("/home/chenxi/tempfile/YOLO_v1/utils/voc2007test.txt") as f:
-        #     lines = f.readlines()
-        # for line in lines:
-        #     splited = line.strip().split()  # ['005246.jpg', '84', '48', '493', '387', '2'] img_name + 坐标 + 类型(labels)
-        #     self.fnames.append(splited[0])
-        #     num_boxes = (len(splited) - 1) // 5
-        #     box = []
-        #     label = []
-        #     for i in range(num_boxes):
-        #         x1 = float(splited[1 + 5 * i]) - 1
-        #         y1 = float(splited[2 + 5 * i]) - 1
-        #         x2 = float(splited[3 + 5 * i]) - 1
-        #         y2 = float(splited[4 + 5 * i]) - 1
-        #         c_label = splited[5 + 5 * i]
-        #         box.append([x1, y1, x2, y2])
-        #         label.append(int(c_label) + 1)
-        #     self.boxes.append(torch.Tensor(box))
-        #     self.labels.append(torch.LongTensor(label))
-        # self.num_samples = len(self.boxes)  # 数据集中包含所有Ground truth个数
-        # =========================================================================
         # _roidb 是一个list, 存放的是每一个xml内部标签,写成字典的格式{"boxes":array([[x1, y1, x2, y2]], "gt_classes":array([[label]]}
         self._roidb = self.gt_roidb()
 
@@ -167,11 +146,11 @@ class RoiDataset(Dataset):
         This function loads/saves from/to a cache file to speed up future calls.
         """
         cache_file = os.path.join(self.cache_path, self.name + '_gt_roidb.pkl')
-        # if os.path.exists(cache_file):
-        #     with open(cache_file, 'rb') as fid:
-        #         roidb = pickle.load(fid)
-        #     print('{} gt roidb loaded from {}'.format(self.name, cache_file))
-        #     return roidb
+        if os.path.exists(cache_file):
+            with open(cache_file, 'rb') as fid:
+                roidb = pickle.load(fid)
+            print('{} gt roidb loaded from {}'.format(self.name, cache_file))
+            return roidb
 
         gt_roidb = [self._load_pascal_annotation(index) for index in self.image_index]
         with open(cache_file, 'wb') as fid:
@@ -247,7 +226,7 @@ def detection_collate(batch):
     # kind of hack, this will break down a list of tuple into
     # individual list
     bsize = len(batch)
-    im_data, boxes, gt_classes, num_obj = zip(*batch)
+    im_data, boxes, gt_classes, num_obj = zip(*batch)  # 只取得最少的那个分配给变量
     max_num_obj = max([x.item() for x in num_obj])
     padded_boxes = torch.zeros((bsize, max_num_obj, 4))
     padded_classes = torch.zeros((bsize, max_num_obj,))
